@@ -1,58 +1,36 @@
 {
-  "name": "next-ecommerce",
+  "name": "startup-nextjs-template",
   "version": "0.1.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "next build",
+    "lint": "next lint",
     "start": "next start",
-    "export": "next build && next export"
+    "build": "next build",
+    "test": "next test",
+    "eject": "next eject"
   },
-  "keywords": [
-    "react",
-    "reactjs",
-    "ecommerce",
-    "nextjs",
-    "next"
-  ],
   "dependencies": {
-    "@heroicons/react": "^2.1.1",
-    "@nextui-org/react": "^2.2.9",
-    "@prisma/client": "^5.7.1",
-    "@reduxjs/toolkit": "^1.8.1",
-    "@tailwindcss/forms": "^0.5.7",
-    "framer-motion": "^10.16.16",
-    "lodash": "^4.17.21",
-    "next": "^12.1.0",
-    "next-redux-wrapper": "^6.0.2",
-    "prisma": "^5.7.1",
-    "rc-slider": "^9.3.1",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
-    "react-hook-form": "^6.8.2",
-    "react-rater": "^5.1.1",
-    "react-redux": "^7.2.1",
-    "react-select": "^3.1.0",
-    "redux": "^4.0.5",
-    "redux-devtools-extension": "^2.13.8",
-    "redux-persist": "^6.0.0",
-    "redux-thunk": "^2.3.0",
-    "reselect": "^4.0.0",
-    "sass": "^1.26.10",
-    "swiper": "^6.5.1",
-    "swr": "^0.3.2",
-    "use-onclickoutside": "^0.3.1"
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mui/material": "^5.14.9",
+    "@next/font": "13.1.6",
+    "eslint": "8.33.0",
+    "eslint-config-next": "13.1.6",
+    "next": "13.1.6",
+    "next-themes": "^0.2.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-modal-video": "^2.0.0"
   },
   "devDependencies": {
-    "@types/lodash": "^4.14.182",
-    "@types/node": "^17.0.34",
-    "@types/react": "^18.0.9",
-    "@types/react-dom": "^18.0.4",
-    "autoprefixer": "^10.4.16",
-    "fslightbox-react": "^1.5.0",
-    "node-fetch": "^3.1.1",
-    "postcss": "^8.4.32",
-    "tailwindcss": "^3.4.0",
-    "typescript": "^4.6.4"
+    "@types/node": "^18.15.11",
+    "@types/react": "^18.0.34",
+    "autoprefixer": "^10.4.13",
+    "postcss": "^8.4.21",
+    "prettier": "^2.8.3",
+    "prettier-plugin-tailwindcss": "^0.2.2",
+    "tailwindcss": "^3.2.4",
+    "typescript": "^5.0.4"
   }
 }
