@@ -1,88 +1,171 @@
-import { useState, useEffect, useRef } from 'react';
-import useOnClickOutside from 'use-onclickoutside';
-import Link from 'next/link';
-import { useRouter } from 'next/router';
-import Image from 'next/image';
+"use client";
+import Image from "next/image";
+import Link from "next/link";
+import { useEffect, useState } from "react";
+// import ThemeToggler from "./ThemeToggler";
+import menuData from "./menuData";
 
-type HeaderType = {
-  isErrorPage?: Boolean;
-}
+const Header = () => {
+  // Navbar toggle
+  const [navbarOpen, setNavbarOpen] = useState(false);
+  const navbarToggleHandler = () => {
+    setNavbarOpen(!navbarOpen);
+  };
 
-const Header = ({ isErrorPage }: HeaderType) => {
-  const router = useRouter();
-  const arrayPaths = ['/'];  
-
-  const [onTop, setOnTop] = useState(( !arrayPaths.includes(router.pathname) || isErrorPage ) ? false : true);
-  const [menuOpen, setMenuOpen] = useState(false);
-  const navRef = useRef(null);
-
-  const headerClass = () => {
-    if(window.pageYOffset === 0) {
-      setOnTop(true);
+  // Sticky Navbar
+  const [sticky, setSticky] = useState(false);
+  const handleStickyNavbar = () => {
+    if (window.scrollY >= 80) {
+      setSticky(true);
     } else {
-      setOnTop(false);
+      setSticky(false);
     }
-  }
-
+  };
   useEffect(() => {
-    if(!arrayPaths.includes(router.pathname) || isErrorPage) {
-      return;
+    window.addEventListener("scroll", handleStickyNavbar);
+  });
+
+  // submenu handler
+  const [openIndex, setOpenIndex] = useState(-1);
+  const handleSubmenu = (index) => {
+    if (openIndex === index) {
+      setOpenIndex(-1);
+    } else {
+      setOpenIndex(index);
     }
+  };
 
-    headerClass();
-    window.onscroll = function() {
-      headerClass();
-    };
-  }, []);
-
-  const closeMenu = () => {
-    setMenuOpen(false);
-  }
-
-
-  // on click outside
-  useOnClickOutside(navRef, closeMenu);
-
-  return(
-    <header className={`site-header ${!onTop ? 'site-header--fixed' : ''}`}>
-      <div className="container">
-        <Link href="/">
-          <a>
-            <Image
-              src="/images/ecowitty_logo.png"
-              width={200}
-              height={100}
-            />
-          </a>
-        </Link>
-        <nav ref={navRef} className={`site-nav ${menuOpen ? 'site-nav--open' : ''}`}>
-          <Link href="/products">
-            <a>Products</a>
-          </Link>
-          <Link href="/ourstory">
-          <a href="#">Our Story</a>
-          </Link>
-          <Link href="/aboutus">
-          <a href="#">About</a>
-          </Link>
-          <Link href="/why">
-          <a href="#">Why Ecowitty</a>
-          </Link>
-          <Link href="/writing">
-          <a href="#">Use</a>
-          </Link>
-        </nav>
-        
-
-        <div className="site-header__actions">
-          <Link href="/contact">
-            <button className={`site-nav ${menuOpen ? 'site-nav--open' : ''}`}>Get in touch</button>
-          </Link>
+  return (
+    <>
+      <header
+        className={`header top-0 left-0 z-40 flex w-full items-center bg-transparent ${
+          sticky
+            ? "!fixed !z-[9999] !bg-white !bg-opacity-80 shadow-sticky backdrop-blur-sm !transition dark:!bg-primary dark:!bg-opacity-20"
+            : "absolute"
+        }`}
+      >
+        <div className="container">
+          <div className="relative flex items-center justify-between">
+            <div className="w-60 max-w-full px-4 xl:mr-12">
+              <Link
+                href="/"
+                className={`header-logo block w-full ${
+                  sticky ? "py-5 lg:py-2" : "py-8"
+                } `}
+              >
+                <Image
+                  src="/images/logo/logo-2.svg"
+                  alt="logo"
+                  width={140}
+                  height={30}
+                  className="w-full dark:hidden"
+                />
+                <Image
+                  src="/images/hero/ecowitty-whiteLogo.png"
+                  alt="logo"
+                  width={140}
+                  height={30}
+                  className="hidden w-full dark:block"
+                />
+              </Link>
+            </div>
+            <div className="flex w-full items-center justify-between px-4">
+              <div>
+                <button
+                  onClick={navbarToggleHandler}
+                  id="navbarToggler"
+                  aria-label="Mobile Menu"
+                  className="absolute right-4 top-1/2 block translate-y-[-50%] rounded-lg px-3 py-[6px] ring-primary focus:ring-2 lg:hidden"
+                >
+                  <span
+                    className={`relative my-1.5 block h-0.5 w-[30px] bg-black transition-all duration-300 dark:bg-white ${
+                      navbarOpen ? " top-[7px] rotate-45" : " "
+                    }`}
+                  />
+                  <span
+                    className={`relative my-1.5 block h-0.5 w-[30px] bg-black transition-all duration-300 dark:bg-white ${
+                      navbarOpen ? "opacity-0 " : " "
+                    }`}
+                  />
+                  <span
+                    className={`relative my-1.5 block h-0.5 w-[30px] bg-black transition-all duration-300 dark:bg-white ${
+                      navbarOpen ? " top-[-8px] -rotate-45" : " "
+                    }`}
+                  />
+                </button>
+                <nav
+                  id="navbarCollapse"
+                  className={`navbar absolute right-0 z-30 w-[250px] rounded border-[.5px] border-body-color/50 bg-white py-4 px-6 duration-300 dark:border-body-color/20 dark:bg-dark lg:visible lg:static lg:w-auto lg:border-none lg:!bg-transparent lg:p-0 lg:opacity-100 ${
+                    navbarOpen
+                      ? "visibility top-full opacity-100"
+                      : "invisible top-[120%] opacity-0"
+                  }`}
+                >
+                  <ul className="block lg:flex lg:space-x-12">
+                    {menuData.map((menuItem, index) => (
+                      <li key={menuItem.id} className="group relative">
+                        {menuItem.path ? (
+                          <Link
+                            href={menuItem.path}
+                            className={`flex py-2 text-base text-dark group-hover:opacity-70 dark:text-white lg:mr-0 lg:inline-flex lg:py-6 lg:px-0`}
+                          >
+                            {menuItem.title}
+                          </Link>
+                        ) : (
+                          <>
+                            <a
+                              onClick={() => handleSubmenu(index)}
+                              className="flex cursor-pointer items-center justify-between py-2 text-base text-dark group-hover:opacity-70 dark:text-white lg:mr-0 lg:inline-flex lg:py-6 lg:px-0"
+                            >
+                              {menuItem.title}
+                              <span className="pl-3">
+                                <svg width="15" height="14" viewBox="0 0 15 14">
+                                  <path
+                                    d="M7.81602 9.97495C7.68477 9.97495 7.57539 9.9312 7.46602 9.8437L2.43477 4.89995C2.23789 4.70308 2.23789 4.39683 2.43477 4.19995C2.63164 4.00308 2.93789 4.00308 3.13477 4.19995L7.81602 8.77183L12.4973 4.1562C12.6941 3.95933 13.0004 3.95933 13.1973 4.1562C13.3941 4.35308 13.3941 4.65933 13.1973 4.8562L8.16601 9.79995C8.05664 9.90933 7.94727 9.97495 7.81602 9.97495Z"
+                                    fill="currentColor"
+                                  />
+                                </svg>
+                              </span>
+                            </a>
+                            <div
+                              className={`submenu relative top-full left-0 rounded-md bg-white transition-[top] duration-300 group-hover:opacity-100 dark:bg-dark lg:invisible lg:absolute lg:top-[110%] lg:block lg:w-[250px] lg:p-4 lg:opacity-0 lg:shadow-lg lg:group-hover:visible lg:group-hover:top-full ${
+                                openIndex === index ? "block" : "hidden"
+                              }`}
+                            >
+                              {menuItem.submenu.map((submenuItem) => (
+                                <Link
+                                  href={submenuItem.path}
+                                  key={submenuItem.id}
+                                  className="block rounded py-2.5 text-sm text-dark hover:opacity-70 dark:text-white lg:px-3"
+                                >
+                                  {submenuItem.title}
+                                </Link>
+                              ))}
+                            </div>
+                          </>
+                        )}
+                      </li>
+                    ))}
+                  </ul>
+                </nav>
+              </div>
+              <div className="flex items-center justify-end pr-16 lg:pr-0">
+                <Link
+                  href="/contact"
+                  className="hidden py-3 px-7 text-base font-bold text-dark hover:opacity-70 dark:text-white md:block"
+                >
+                  Contact Us
+                </Link>
+                {/* <div>
+                  <ThemeToggler />
+                </div> */}
+              </div>
+            </div>
+          </div>
         </div>
-      </div>
-    </header>
-  )
+      </header>
+    </>
+  );
 };
-
 
 export default Header;
