@@ -1,7 +1,8 @@
-const path = require('path')
-
-module.exports = {
-  sassOptions: {
-    includePaths: [path.join(__dirname, 'styles')],
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  experimental: {
+    appDir: true,
   },
-}
+};
+
+module.exports = nextConfig;
